extension gallery manifest service - send gallery info from workbench to shared process
