+++ src/vs/platform/extensionManagement/common/extensionGalleryManifest.ts
@@ -1,0 +1,69 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Event } from '../../../base/common/event.js';
+import { createDecorator } from '../../instantiation/common/instantiation.js';
+
+export const enum ExtensionGalleryResourceType {
+	ExtensionQueryService = 'ExtensionQueryService',
+	ExtensionLatestVersionUri = 'ExtensionLatestVersionUriTemplate',
+	ExtensionStatisticsUri = 'ExtensionStatisticsUriTemplate',
+	WebExtensionStatisticsUri = 'WebExtensionStatisticsUriTemplate',
+	PublisherViewUri = 'PublisherViewUriTemplate',
+	ExtensionDetailsViewUri = 'ExtensionDetailsViewUriTemplate',
+	ExtensionRatingViewUri = 'ExtensionRatingViewUriTemplate',
+}
+
+export const enum Flag {
+	None = 'None',
+	IncludeVersions = 'IncludeVersions',
+	IncludeFiles = 'IncludeFiles',
+	IncludeCategoryAndTags = 'IncludeCategoryAndTags',
+	IncludeSharedAccounts = 'IncludeSharedAccounts',
+	IncludeVersionProperties = 'IncludeVersionProperties',
+	ExcludeNonValidated = 'ExcludeNonValidated',
+	IncludeInstallationTargets = 'IncludeInstallationTargets',
+	IncludeAssetUri = 'IncludeAssetUri',
+	IncludeStatistics = 'IncludeStatistics',
+	IncludeLatestVersionOnly = 'IncludeLatestVersionOnly',
+	Unpublished = 'Unpublished',
+	IncludeNameConflictInfo = 'IncludeNameConflictInfo',
+	IncludeLatestPrereleaseAndStableVersionOnly = 'IncludeLatestPrereleaseAndStableVersionOnly',
+}
+
+export type ExtensionGalleryManifestResource = {
+	readonly id: string;
+	readonly type: string;
+};
+
+export type ExtensionQueryCapabilityValue = {
+	readonly name: string;
+	readonly value: number;
+};
+
+export interface IExtensionGalleryManifest {
+	readonly version: string;
+	readonly resources: readonly ExtensionGalleryManifestResource[];
+	readonly capabilities: {
+		readonly extensionQuery: {
+			readonly filtering?: readonly ExtensionQueryCapabilityValue[];
+			readonly sorting?: readonly ExtensionQueryCapabilityValue[];
+			readonly flags?: readonly ExtensionQueryCapabilityValue[];
+		};
+		readonly signing?: {
+			readonly allRepositorySigned: boolean;
+		};
+	};
+}
+
+export const IExtensionGalleryManifestService = createDecorator<IExtensionGalleryManifestService>('IExtensionGalleryManifestService');
+
+export interface IExtensionGalleryManifestService {
+	readonly _serviceBrand: undefined;
+
+	readonly onDidChangeExtensionGalleryManifest: Event<IExtensionGalleryManifest | null>;
+	isEnabled(): boolean;
+	getExtensionGalleryManifest(): Promise<IExtensionGalleryManifest | null>;
+}
